requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.41.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
aiomysql>=0.3.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
aiofiles>=25.0.0
jinja2>=3.1.0
